requirements.txt
asttokens==2.0.8
async-generator==1.10
attrs==22.1.0
autopep8==1.7.0
backcall==0.2.0
beautifulsoup4==4.11.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
debugpy==1.6.3
decorator==5.1.1
distlib==0.3.6
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.0.0rc9
executing==1.1.1
filelock==3.8.0
h11==0.14.0
idna==3.4
ipykernel==6.16.0
ipython==8.5.0
jedi==0.18.1
jupyter-core==4.11.1
jupyter_client==7.4.2
matplotlib-inline==0.1.6
MouseInfo==0.1.3
nest-asyncio==1.5.6
numpy==1.23.4
opencv-python==4.6.0.66
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pandas==1.5.0
parso==0.8.3
pdf2image==1.16.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.3
prompt-toolkit==3.0.31
psutil==5.9.2
pure-eval==0.2.2
PyAutoGUI==0.9.53
pycodestyle==2.9.1
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.13.0
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytweening==1.0.4
pytz==2022.4
pywin32==304
pyzbar==0.1.9
pyzmq==24.0.1
requests==2.28.1
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.5.1
toml==0.10.2
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
virtualenv==20.16.6
wcwidth==0.2.5
webdriver-manager==3.8.3
wsproto==1.2.0
